# daemon/openapi3.yaml
openapi: "3.0.0"

info:
  version: 0.0.1
  title: Steady Daemon

paths:
  /steady/application/{name}:
    parameters:
      - name: name
        in: path
        description: tags to filter by
        required: true
        schema:
          type: string
    post:
      operationId: createApplication
      requestBody:
        description: Application to add to daemon
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - script
              properties:
                script:
                  type: string
      responses:
        "201":
          description: "success response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: getApplication
      responses:
        "200":
          description: "application"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteApplication
      responses:
        "200":
          description: "application"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Application:
      type: object
      required:
        - name
        - start_count
        - request_count
      properties:
        name:
          type: string
        start_count:
          type: integer
        request_count:
          type: integer

    Error:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string